requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
jinja2==3.1.4
pydantic==2.9.2
python-multipart==0.0.12
requests==2.32.3
httpx==0.27.2
pytesseract==0.3.13
Pillow==10.4.0
opencv-python-headless==4.10.0.84
numpy==2.1.2
APScheduler==3.10.4
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
tenacity==9.0.0
loguru==0.7.2
